sdk: improved capture read test
